ticket-history/pyproject.toml
[project]
name = "ticket-history"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "httpx",
    "python-dotenv",
    "tqdm",
    "streamlit",
    "plotly",
    "pandas",
]
